requirements.txt
alabaster==0.7.3
amqp==1.4.9
anyjson==0.3.3
arrow==0.8.0
astroid==1.3.6
Babel==1.3
beautifulsoup4==4.3.2
billiard==3.3.0.23
celery==3.1.23
census==0.7
coverage==4.0a5
dbfpy==2.3.1
dj-database-url==0.4.1
django-bootstrap-form==3.2
django-bootstrap-forms==0.1
django-bower==5.0.2
django-cors-headers==1.0.0
django-crispy-forms==1.6.0
django-geojson==2.9.0
django-leaflet==0.18.0
django-modalview==0.1.5
django-nose==1.3
djangorestframework==3.1.1
docutils==0.12
geojson==1.3.2
googlemaps==2.4.2
gunicorn==19.6.0
httplib2==0.9
httpretty==0.8.8
jellyfish==0.5.1
Jinja2==2.7.3
jsonfield==1.0.3
kombu==3.0.35
logilab-common==0.63.2
lxml==3.6.0
MarkupSafe==0.23
mock==1.0.1
nose==1.3.4
numpy==1.10.4
oauth2==1.5.211
oauthlib==0.7.2
pdfminer==20140328
PIL==1.1.7
polyline==1.1
Pygments==2.0.2
pylint==1.4.3
pyPdf==1.13
PyPDF2==1.26.0
PySAL==1.11.1
pyshp==1.2.3
python-dateutil==2.5.3
python-google-places==1.2.0
python-openid==2.2.5
PyTumblr==0.0.6
pytz==2015.2
redis==2.10.5
requests==2.6.0
requests-oauthlib==0.4.2
scipy==0.17.0
simplejson==3.6.5
six==1.8.0
snowballstemmer==1.2.0
Sphinx==1.3.1
sphinx-rtd-theme==0.1.7
spotipy==2.3.8
twilio==3.8.0
us==0.9.1
virtualenv==15.0.2
yelp==1.0.2
